requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
PyMuPDF
# chromadb
faiss-cpu
langchain_google_genai==1.0.3
python-dotenv
streamlit_mic_recorder

# pip install streamlit 